Move DigraphResult to top level class
